jwt_fastapi_app/requirements.txt
fastapi
uvicorn
passlib[bcrypt]
PyJWT
httpx
pytest
pytest-asyncio
